implement new API for ElegyModule
