+++ elegy/generalized_module/elegy_module_test.py
@@ -75,7 +75,11 @@ class ElegyModuleTest(TestCase):
 
         model = elegy.Model(ModuleA())
 
-        summary_text = model.summary(x=jnp.ones([10, 2]), depth=1, return_repr=True)
+        x = jnp.ones([10, 2])
+
+        model.init(x=x)
+
+        summary_text = model.summary(x, depth=1, return_repr=True)
         assert summary_text is not None
 
         lines = summary_text.split("\n")
